+++ CHANGELOG.md
@@ -4,15 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.3] - 2022-08-09
 ### Added
 - Swagger UI /swagger/index.html
 - date and time now is showing in UTC
 - changed primary color for the theme
 - added pagination to the pages
 - other minor improvements
-### Changed
-- AppBar color
 
 ## [0.1.2] - 2022-08-02
 ### Changed